Listar evento por categoria

+++ src/main/java/br/com/fatesg/eventos/persistence/EventoPersistence.java
@@ -10,5 +10,5 @@ import br.com.fatesg.eventos.model.Evento;
 
 public interface EventoPersistence extends JpaRepository<Evento, Integer> {
 	@Query("SELECT u from Evento u where u.categoria.codigo = :codigo")
-	public List<Evento> listarEventosPorCategoria(@Param("codigo") Long idCategoria);
+	public List<Evento> listarEventosPorCategoria(@Param("codigo") Integer idCategoria);
 }
